# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/apps": {
            "get": {
                "description": "租户列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户管理"
                ],
                "summary": "租户列表",
                "operationId": "/apps",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "info",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "每页多少条",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page_no",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/public.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.APPListOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "租户更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户管理"
                ],
                "summary": "租户更新",
                "operationId": "/apps/update",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.APPUpdateHttpInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/public.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "租户添加",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户管理"
                ],
                "summary": "租户添加",
                "operationId": "/apps/add",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.APPAddHttpInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/public.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/apps/{id}": {
            "get": {
                "description": "租户详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户管理"
                ],
                "summary": "租户详情",
                "operationId": "/apps/detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/public.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dao.App"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "租户删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户管理"
                ],
                "summary": "租户删除",
                "operationId": "/apps/delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/public.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/apps/{id}/stat": {
            "get": {
                "description": "租户统计",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "租户管理"
                ],
                "summary": "租户统计",
                "operationId": "/apps/stat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/public.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.StatisticsOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/services/{id}/stat": {
            "get": {
                "description": "服务统计",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务管理"
                ],
                "summary": "服务统计",
                "operationId": "/services/:id/stat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/public.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ServiceStatOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dao.App": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "create_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_delete": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "qpd": {
                    "type": "integer"
                },
                "qps": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                },
                "white_ips": {
                    "type": "string"
                }
            }
        },
...
        "dto.APPAddHttpInput": {
            "type": "object",
            "required": [
                "app_id",
                "name"
            ],
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "qpd": {
                    "type": "integer"
                },
                "qps": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                },
                "white_ips": {
                    "type": "string"
                }
            }
        },
        "dto.APPListItemOutput": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "create_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_delete": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "qpd": {
                    "type": "integer"
                },
                "qps": {
                    "type": "integer"
                },
                "real_qpd": {
                    "type": "integer"
                },
                "real_qps": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                },
                "white_ips": {
                    "type": "string"
                }
            }
        },
        "dto.APPListOutput": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.APPListItemOutput"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.APPUpdateHttpInput": {
            "type": "object",
            "required": [
                "id",
                "name",
                "secret"
            ],
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "qpd": {
                    "type": "integer"
                },
                "qps": {
                    "type": "integer"
                },
                "secret": {
                    "type": "string"
                },
                "white_ips": {
                    "type": "string"
                }
            }
        },
...
        "dto.ServiceStatOutput": {
            "type": "object",
            "properties": {
                "today": {
                    "description": "列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "yesterday": {
                    "description": "列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "dto.StatisticsOutput": {
            "type": "object",
            "required": [
                "today",
                "yesterday"
            ],
            "properties": {
                "today": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "yesterday": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "public.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                },
                "stack": {
                    "type": "object"
                },
                "trace_id": {
                    "type": "object"
                }
            }
        }
    }
}